Add initial portfolio structure and style
A rough but the complete version of John Doe's portfolio is introduced. It includes the index.html file for structuring the content of the portfolio, the style.css file for styling the website, a script.js file to implement basic functionality such as smooth scrolling when navigation links are clicked, the .gitignore file inside the .idea directory to ignore file paths of no relevance, the vcs.xml and untitled.iml files in the .idea directory to work with IntelliJ Idea as the IDE and the modules.xml file in the same directory for managing other modules within the project. For the above mentioned files to be tracked by Git, a new file mode, 100644 is set.

The index.html contains structure for sections such as about, experiences, formation, compétences, langues and loisirs while being explanatory internally. The global styles for the website has been added in the styles.css file including media queries for responsiveness. The portfolio is currently displayed in a fixed-section layout for screen sizes of 1251px or more and changes to a more adaptive layout for smaller screens.